+++ mask_rcnn_pytorch/roialign/cuda/crop_and_resize_kernel.cu
@@ -1,6 +1,5 @@
-#include <cuda.h>
-#include <cmath>
-#include <iostream>
+#include <math.h>
+#include <stdio.h>
 #include "crop_and_resize_kernel.h"
 
 #define CUDA_1D_KERNEL_LOOP(i, n)                            \
@@ -10,33 +9,33 @@
 __global__ void CropAndResizeKernel(const int nthreads,
                                     const float* image_ptr,
                                     const float* boxes_ptr,
-                                    const int64_t* box_ind_ptr,
-                                    int64_t num_boxes,
-                                    int64_t batch,
-                                    int64_t image_height,
-                                    int64_t image_width,
-                                    int64_t crop_height,
-                                    int64_t crop_width,
-                                    int64_t depth,
+                                    const int* box_ind_ptr,
+                                    int num_boxes,
+                                    int batch,
+                                    int image_height,
+                                    int image_width,
+                                    int crop_height,
+                                    int crop_width,
+                                    int depth,
                                     float extrapolation_value,
                                     float* crops_ptr) {
   CUDA_1D_KERNEL_LOOP(out_idx, nthreads) {
     // NHWC: out_idx = d + depth * (w + crop_width * (h + crop_height * b))
     // NCHW: out_idx = w + crop_width * (h + crop_height * (d + depth * b))
-    int64_t idx = out_idx;
-    const int64_t x = idx % crop_width;
+    int idx = out_idx;
+    const int x = idx % crop_width;
     idx /= crop_width;
-    const int64_t y = idx % crop_height;
+    const int y = idx % crop_height;
     idx /= crop_height;
-    const int64_t d = idx % depth;
-    const int64_t b = idx / depth;
+    const int d = idx % depth;
+    const int b = idx / depth;
 
     const float y1 = boxes_ptr[b * 4];
     const float x1 = boxes_ptr[b * 4 + 1];
     const float y2 = boxes_ptr[b * 4 + 2];
     const float x2 = boxes_ptr[b * 4 + 3];
 
-    const int64_t b_in = box_ind_ptr[b];
+    const int b_in = box_ind_ptr[b];
     if (b_in < 0 || b_in >= batch) {
       continue;
     }
@@ -49,7 +48,7 @@ __global__ void CropAndResizeKernel(const int nthreads,
 
     const float in_y = (crop_height > 1)
                            ? y1 * (image_height - 1) + y * height_scale
-                           : 0.5f * (y1 + y2) * (image_height - 1);
+                           : 0.5 * (y1 + y2) * (image_height - 1);
     if (in_y < 0 || in_y > image_height - 1) {
       crops_ptr[out_idx] = extrapolation_value;
       continue;
@@ -57,18 +56,18 @@ __global__ void CropAndResizeKernel(const int nthreads,
 
     const float in_x = (crop_width > 1)
                            ? x1 * (image_width - 1) + x * width_scale
-                           : 0.5f * (x1 + x2) * (image_width - 1);
+                           : 0.5 * (x1 + x2) * (image_width - 1);
     if (in_x < 0 || in_x > image_width - 1) {
       crops_ptr[out_idx] = extrapolation_value;
       continue;
     }
 
-    const int64_t top_y_index = static_cast<int64_t>(floorf(in_y));
-    const int64_t bottom_y_index = static_cast<int64_t>(ceilf(in_y));
+    const int top_y_index = floorf(in_y);
+    const int bottom_y_index = ceilf(in_y);
     const float y_lerp = in_y - top_y_index;
 
-    const int64_t left_x_index = floorf(in_x);
-    const int64_t right_x_index = ceilf(in_x);
+    const int left_x_index = floorf(in_x);
+    const int right_x_index = ceilf(in_x);
     const float x_lerp = in_x - left_x_index;
 
     const float* pimage =
@@ -89,32 +88,32 @@ __global__ void CropAndResizeKernel(const int nthreads,
 __global__ void CropAndResizeBackpropImageKernel(const int nthreads,
                                                  const float* grads_ptr,
                                                  const float* boxes_ptr,
-                                                 const int64_t* box_ind_ptr,
-                                                 int64_t num_boxes,
-                                                 int64_t batch,
-                                                 int64_t image_height,
-                                                 int64_t image_width,
-                                                 int64_t crop_height,
-                                                 int64_t crop_width,
-                                                 int64_t depth,
+                                                 const int* box_ind_ptr,
+                                                 int num_boxes,
+                                                 int batch,
+                                                 int image_height,
+                                                 int image_width,
+                                                 int crop_height,
+                                                 int crop_width,
+                                                 int depth,
                                                  float* grads_image_ptr) {
   CUDA_1D_KERNEL_LOOP(out_idx, nthreads) {
     // NHWC: out_idx = d + depth * (w + crop_width * (h + crop_height * b))
     // NCHW: out_idx = w + crop_width * (h + crop_height * (d + depth * b))
     int idx = out_idx;
-    const int64_t x = idx % crop_width;
+    const int x = idx % crop_width;
     idx /= crop_width;
-    const int64_t y = idx % crop_height;
+    const int y = idx % crop_height;
     idx /= crop_height;
-    const int64_t d = idx % depth;
-    const int64_t b = idx / depth;
+    const int d = idx % depth;
+    const int b = idx / depth;
 
     const float y1 = boxes_ptr[b * 4];
     const float x1 = boxes_ptr[b * 4 + 1];
     const float y2 = boxes_ptr[b * 4 + 2];
     const float x2 = boxes_ptr[b * 4 + 3];
 
-    const int64_t b_in = box_ind_ptr[b];
+    const int b_in = box_ind_ptr[b];
     if (b_in < 0 || b_in >= batch) {
       continue;
     }
@@ -127,24 +126,24 @@ __global__ void CropAndResizeBackpropImageKernel(const int nthreads,
 
     const float in_y = (crop_height > 1)
                            ? y1 * (image_height - 1) + y * height_scale
-                           : 0.5f * (y1 + y2) * (image_height - 1);
+                           : 0.5 * (y1 + y2) * (image_height - 1);
     if (in_y < 0 || in_y > image_height - 1) {
       continue;
     }
 
     const float in_x = (crop_width > 1)
                            ? x1 * (image_width - 1) + x * width_scale
-                           : 0.5f * (x1 + x2) * (image_width - 1);
+                           : 0.5 * (x1 + x2) * (image_width - 1);
     if (in_x < 0 || in_x > image_width - 1) {
       continue;
     }
 
-    const int64_t top_y_index = static_cast<int64_t>(floorf(in_y));
-    const int64_t bottom_y_index = static_cast<int64_t>(ceilf(in_y));
+    const int top_y_index = floorf(in_y);
+    const int bottom_y_index = ceilf(in_y);
     const float y_lerp = in_y - top_y_index;
 
-    const int64_t left_x_index = static_cast<int64_t>(floorf(in_x));
-    const int64_t right_x_index = static_cast<int64_t>(ceilf(in_x));
+    const int left_x_index = floorf(in_x);
+    const int right_x_index = ceilf(in_x);
     const float x_lerp = in_x - left_x_index;
 
     float* pimage =
@@ -165,19 +164,19 @@ __global__ void CropAndResizeBackpropImageKernel(const int nthreads,
 
 void CropAndResizeLaucher(const float* image_ptr,
                           const float* boxes_ptr,
-                          const int64_t* box_ind_ptr,
-                          int64_t num_boxes,
-                          int64_t batch,
-                          int64_t image_height,
-                          int64_t image_width,
-                          int64_t crop_height,
-                          int64_t crop_width,
-                          int64_t depth,
+                          const int* box_ind_ptr,
+                          int num_boxes,
+                          int batch,
+                          int image_height,
+                          int image_width,
+                          int crop_height,
+                          int crop_width,
+                          int depth,
                           float extrapolation_value,
                           float* crops_ptr) {
-  const int64_t total_count = num_boxes * crop_height * crop_width * depth;
-  const int64_t thread_per_block = 1024;
-  const int64_t block_count =
+  const int total_count = num_boxes * crop_height * crop_width * depth;
+  const int thread_per_block = 1024;
+  const int block_count =
       (total_count + thread_per_block - 1) / thread_per_block;
   cudaError_t err;
 
@@ -198,18 +197,18 @@ void CropAndResizeLaucher(const float* image_ptr,
 
 void CropAndResizeBackpropImageLaucher(const float* grads_ptr,
                                        const float* boxes_ptr,
-                                       const int64_t* box_ind_ptr,
-                                       int64_t num_boxes,
-                                       int64_t batch,
-                                       int64_t image_height,
-                                       int64_t image_width,
-                                       int64_t crop_height,
-                                       int64_t crop_width,
-                                       int64_t depth,
+                                       const int* box_ind_ptr,
+                                       int num_boxes,
+                                       int batch,
+                                       int image_height,
+                                       int image_width,
+                                       int crop_height,
+                                       int crop_width,
+                                       int depth,
                                        float* grads_image_ptr) {
-  const int64_t total_count = num_boxes * crop_height * crop_width * depth;
-  const int64_t thread_per_block = 1024;
-  const int64_t block_count =
+  const int total_count = num_boxes * crop_height * crop_width * depth;
+  const int thread_per_block = 1024;
+  const int block_count =
       (total_count + thread_per_block - 1) / thread_per_block;
   cudaError_t err;
